Add files via upload
current problem: google docs keeps rate limiting me for trying to pull from the actual document. The goal is to have a editor for the volunteers who don't know how to use markdown and then convert it into an equivalent markdown file to then render on the website. 

Idea: you can export a google doc as markdown so maybe the workflow could be to make the appropriate changes in a simple editor of our choice, export to markdown and upload it to github, then the google apps script can call the github api (which has less limits) to get the markdown instead.